requirements.txt
mediapipe>=0.10.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
gradio>=4.0.0
pytest>=7.0.0

